Add scoverage plugin + cleanup code & tests
